requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pinecone-client
openai
networkx
spacy
python-multipart
aiofiles
tqdm
requests
pdfplumber